feat(delivery,ops): transactional outbox, suppression, webhooks, Docker stack, hardened CI
Phase 5 plus the operational scaffolding. Addresses C-04, C-05, C-06, C-11,
C-13, H-17, H-18, M-12.

Delivery. The outbox is now the single chokepoint for every outbound message:
  - lease via FOR UPDATE SKIP LOCKED, so N workers divide the queue safely
  - the FULL authorization chain is re-evaluated at the send boundary, not
    trusted from when the row was queued; pausing a campaign, a reply, or an
    unsubscribe stops mail that is already sitting in the queue
  - quota reserved atomically in the same transaction
  - provider idempotency key stored before the first attempt, so a crash after
    the provider accepted does not duplicate on retry
  - quota exhaustion DEFERS with deterministic spread, never fails permanently
  - a permanent recipient rejection suppresses; a configuration error does not

Webhooks. Duplicates collapse on UNIQUE(provider, event_id); state advances by
rank so a delayed 'sent' cannot overwrite 'bounced' and a late 'opened' cannot
mask a complaint; hard bounces and complaints suppress, soft bounces do not.
Resend signature verification is Svix-style, constant-time, replay-bounded, and
fails closed when no secret is configured.

Two real bugs the tests caught and this fixes:
  - suppressing a plus-tagged address left the base address reachable; the
    base is now recorded at write time
  - the domain-scope suppression query needed explicit grouping

Operations:
  - docker-compose rebuilt: every service that needs a setting receives it, so
    the 'configured in .env but invisible to the runtime' defect cannot recur.
    Separate Temporal database, real temporal-ui image (the old file published
    8233 on a gRPC-only container), healthchecks, non-root, read-only rootfs.
    The Temporal worker service is deliberately absent because Phase 7 is not
    implemented -- a service pointing at a nonexistent module would be fiction.
  - .env.example is now GENERATED from the Settings model (75 vars); CI diffs it
  - CI has no '|| true' anywhere; adds secret scanning, dependency audit,
    migration round trip, schema drift, coverage floor on safety-critical
    modules, compose assertions that the kill switch is off, and a non-root
    image check. Actions pinned to commit SHAs.
  - titan CLI: preflight, check-providers, env-example, invariants
  - API ships health/ready/sending-preflight only; no stub routes returning
    plausible empty data

24 repository invariant tests enforce section 28 statically (AST scan).

Verified: 274 tests pass; ruff clean; docker compose config valid.

+++ apps/api/titan/intelligence/scoring.py
@@ -294,7 +294,7 @@ def score_lead(data: ScoringInput, threshold: int = 70) -> ScoreResult:
         penalty += OUTREACH_RISK_PENALTY
         reasons.append("Outreach risk: " + ", ".join(data.outreach_risk_flags))
 
-    total = int(round(max(0.0, min(100.0, subtotal - penalty))))
+    total = round(max(0.0, min(100.0, subtotal - penalty)))
 
     # ---- hard gates -------------------------------------------------------
     # A lead with nothing evidenced cannot be "qualified" regardless of fit:
